requirements.txt
pyautogui>=0.9.54
pyinstaller>=5.0
pillow>=9.0.0 